ファイルロード機能をModel側から分離
